requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
transformers>=4.30.0
peft>=0.4.0
wandb>=0.15.0
tensorboard>=2.13.0
numpy>=1.24.0
tqdm>=4.65.0
pytorch-hessian-eigenthings>=0.1.0

